Restrict thread deletion to thread poster
