+++ frontend/src/App.tsx
@@ -4,7 +4,7 @@ import Login from "./components/Login";
 import Register from "./components/Register";
 import Reset from "./components/Reset";
 import Menu from "./components/Menu"
-import CreatePost from "./components/CreatePost";
+import CreateThread from "./components/CreateThread";
 
 const App: React.FC = () => {
   return (
@@ -14,7 +14,7 @@ const App: React.FC = () => {
           <Route path="/register" element={<Register />} />
           <Route path="/reset" element={<Reset />} />
           <Route path="/menu" element={<Menu />} />
-          <Route path="/create" element={<CreatePost />} />
+          <Route path="/create" element={<CreateThread />} />
         </Routes>
       </div>
   );
